cloud-functions/get-nhl-json/requirements.txt
requests==2.24.0
google-cloud-storage==1.31.0
simplejson==3.17.2